Move data dir funcs to config
