+++ src/config/index.js
@@ -1,6 +1,6 @@
-const fs = require('fs');
-const { includes, isNumber } = require('lodash');
-
+const fs = require('fs-extra');
+const path = require('path');
+const { isEmpty, includes, isNumber } = require('lodash');
 const { BLOCKCHAIN_ENV } = require('../constants');
 const contractMetadata = require('./contract-metadata');
 
@@ -15,6 +15,44 @@ const CONFIG = {
   HOSTNAME: 'localhost',
   PROTOCOL: includes(process.argv, '--local') ? 'http' : 'https',
 };
+
+/**
+ * Returns the base data dir path, and also creates it if necessary.
+ * @return {string} Path to the base data directory.
+ */
+const getBaseDataDir = () => {
+  // DATA_DIR is defined in environment variables
+  if (!isEmpty(process.env.DATA_DIR)) {
+    return process.env.DATA_DIR;
+  }
+
+  const rootDir = path.resolve('../../');
+  const dataDir = `${rootDir}/data/${CONFIG.NETWORK}`;
+  fs.ensureDirSync(dataDir);
+  return path.resolve(dataDir);
+};
+
+/**
+ * Returns the database dir path, and also creates it if necesssary.
+ * @return {string} Path to the database directory.
+ */
+const getDbDir = () => {
+  const baseDir = getBaseDataDir();
+  const dbDir = `${baseDir}/nedb`;
+  fs.ensureDirSync(dbDir);
+  return path.resolve(dbDir);
+};
+
+/**
+ * Returns the logs dir path, and also creates it if necesssary.
+ * @return {string} Path to the logs directory.
+ */
+function getLogsDir() {
+  const basePath = getBaseDataDir();
+  const logsDir = `${basePath}/logs`;
+  fs.ensureDirSync(path);
+  return path.resolve(logsDir);
+}
 
 const isMainnet = () => CONFIG.NETWORK === BLOCKCHAIN_ENV.MAINNET;
 
@@ -49,6 +87,8 @@ const getSSLCredentials = () => {
 
 module.exports = {
   CONFIG,
+  getDbDir,
+  getLogsDir,
   isMainnet,
   getContractMetadata,
   getContractAddress,
